Separate out parser-specific pytest unit testing.
- Discovery and parsing are related but belong in separate files
