+++ src/test/unittests/pytest/pytest.discovery.unit.test.ts
@@ -9,19 +9,15 @@ import { expect, use } from 'chai';
 import * as chaipromise from 'chai-as-promised';
 import * as path from 'path';
 import * as typeMoq from 'typemoq';
-import { CancellationToken, OutputChannel, Uri } from 'vscode';
-import { getOSType } from '../../..//client/common/platform/osinfo';
-import { IApplicationShell, ICommandManager } from '../../../client/common/application/types';
-import { OSType } from '../../../client/common/platform/types';
+import { CancellationToken } from 'vscode';
 import { IServiceContainer } from '../../../client/ioc/types';
 import { PYTEST_PROVIDER } from '../../../client/unittests/common/constants';
-import { TestsHelper } from '../../../client/unittests/common/testUtils';
-import { TestFlatteningVisitor } from '../../../client/unittests/common/testVisitors/flatteningVisitor';
-import { FlattenedTestFunction, ITestDiscoveryService, ITestRunner, ITestsHelper, ITestsParser, Options, TestDiscoveryOptions, Tests } from '../../../client/unittests/common/types';
+import {
+    ITestDiscoveryService, ITestRunner, ITestsHelper,
+    ITestsParser, Options, TestDiscoveryOptions, Tests
+} from '../../../client/unittests/common/types';
 import { TestDiscoveryService } from '../../../client/unittests/pytest/services/discoveryService';
-import { TestsParser as PyTestsParser } from '../../../client/unittests/pytest/services/parserService';
 import { IArgumentsService, TestFilter } from '../../../client/unittests/types';
-import { PytestDataPlatformType, pytestScenarioData } from './pytest_unittest_parser_data';
 
 use(chaipromise);
 
@@ -185,59 +181,4 @@ suite('Unit Tests - PyTest - Discovery', () => {
         runner.verifyAll();
         testParser.verifyAll();
     });
-    // build tests for the test data that is relevant for this platform.
-    pytestScenarioData.forEach((testScenario) => {
-        const testPlatformType: PytestDataPlatformType =
-            getOSType() === OSType.Windows ?
-                PytestDataPlatformType.Windows : PytestDataPlatformType.NonWindows;
-
-        // tslint:disable-next-line:no-invalid-this
-        if (testPlatformType === testScenario.platform) {
-            const testDescription: string = 'PyTest '.concat('[', testScenario.pytest_version_spec, '] ', testScenario.description);
-            test(testDescription, async () => {
-                const testHelper = typeMoq.Mock.ofType<TestsHelper>();
-                testHelper.setup(h => h.flattenTestFiles(typeMoq.It.isAny()))
-                    .returns((v) => v);
-                const testFlattener: TestFlatteningVisitor = new TestFlatteningVisitor();
-                const serviceContainer = typeMoq.Mock.ofType<IServiceContainer>();
-                const appShell = typeMoq.Mock.ofType<IApplicationShell>();
-                const cmdMgr = typeMoq.Mock.ofType<ICommandManager>();
-                serviceContainer.setup(s => s.get(typeMoq.It.isValue(IApplicationShell), typeMoq.It.isAny()))
-                    .returns(() => {
-                        return appShell.object;
-                    });
-                serviceContainer.setup(s => s.get(typeMoq.It.isValue(ICommandManager), typeMoq.It.isAny()))
-                    .returns(() => {
-                        return cmdMgr.object;
-                    });
-                const forRealzTestHelper: TestsHelper = new TestsHelper(testFlattener, serviceContainer.object);
-                const parser = new PyTestsParser(forRealzTestHelper);
-                const outChannel = typeMoq.Mock.ofType<OutputChannel>();
-                const cancelToken = typeMoq.Mock.ofType<CancellationToken>();
-                cancelToken.setup(c => c.isCancellationRequested).returns(() => false);
-                const wsFolder = typeMoq.Mock.ofType<Uri>();
-
-                const options: TestDiscoveryOptions = {
-                    args: [],
-                    cwd: testScenario.rootdir,
-                    ignoreCache: true,
-                    outChannel: outChannel.object,
-                    token: cancelToken.object,
-                    workspaceFolder: wsFolder.object
-                };
-                const parsedTests: Tests = parser.parse(testScenario.stdout.join('\n'), options);
-                expect(parsedTests).is.not.equal(undefined, 'Should have gotten tests extracted from the parsed pytest result content.');
-                expect(parsedTests.testFunctions.length).equals(testScenario.functionCount, `Parsed pytest summary contained ${testScenario.functionCount} test functions.`);
-                testScenario.test_functions.forEach((funcName: string) => {
-                    const findOneTest: FlattenedTestFunction | undefined = parsedTests.testFunctions.find(
-                        (tstFunc: FlattenedTestFunction) => {
-                            return tstFunc.testFunction.nameToRun === funcName;
-                        });
-                    expect(findOneTest).is.not.equal(undefined, `Could not find "${funcName}" in tests.`);
-                });
-
-            });
-        }
-    });
-
 });
